handle no connection

+++ app/src/main/java/app/rootstock/ui/messages/MessagesViewModel.kt
@@ -15,15 +15,17 @@ import app.rootstock.data.messages.Message
 import app.rootstock.data.messages.MessageRepository
 import app.rootstock.data.network.ResponseResult
 import app.rootstock.data.result.Event
+import app.rootstock.utils.InternetUtil
 import dagger.hilt.android.scopes.ActivityScoped
 import kotlinx.coroutines.flow.Flow
 import kotlinx.coroutines.flow.first
 import kotlinx.coroutines.launch
 
-sealed class MessageEventS() {
-    class Error(val message: String) : MessageEventS()
-    class Created() : MessageEventS()
-    class Deleted() : MessageEventS()
+sealed class MessageEvent() {
+    class Error(val message: String) : MessageEvent()
+    class Created() : MessageEvent()
+    class Deleted() : MessageEvent()
+    class NoConnection(val attempt: String? = null) : MessageEvent()
 }
 
 @ActivityScoped
@@ -31,8 +33,8 @@ sealed class MessageEventS() {
 class MessagesViewModel @ViewModelInject constructor(private val repository: MessageRepository) :
     ViewModel() {
 
-    private val _messageEvent = MutableLiveData<Event<MessageEventS>>()
-    val messageEvent: LiveData<Event<MessageEventS>> get() = _messageEvent
+    private val _messageEvent = MutableLiveData<Event<MessageEvent>>()
+    val messageEvent: LiveData<Event<MessageEvent>> get() = _messageEvent
 
     private val _channel = MutableLiveData<Channel>()
     val channel: LiveData<Channel>
@@ -61,39 +63,52 @@ class MessagesViewModel @ViewModelInject constructor(private val repository: Mes
         return newResult
     }
 
+    private fun isConnected(content: String? = null): Boolean {
+        if (!InternetUtil.isInternetOn()) {
+            _messageEvent.value = Event(MessageEvent.NoConnection(content))
+            return false
+        }
+        return true
+    }
+
     fun sendMessage(content: String?) {
         content ?: return
         val id = _channel.value?.channelId ?: return
+        if (!isConnected(content)) return
 
         val sendMessage = SendMessage(content = content, channelId = id)
+        sendMessage(sendMessage)
+    }
 
+    private fun sendMessage(sendMessage: SendMessage) {
         viewModelScope.launch {
             val wsId = channel.value?.workspaceId ?: return@launch
             when (val message =
                 repository.sendMessage(message = sendMessage, workspaceId = wsId).first()) {
                 is ResponseResult.Success -> {
-                    _messageEvent.value = Event(MessageEventS.Created())
+                    _messageEvent.value = Event(MessageEvent.Created())
                     modifyChannel()
                 }
                 is ResponseResult.Error -> {
-                    _messageEvent.postValue(Event(MessageEventS.Error(message = message.message)))
+                    _messageEvent.postValue(Event(MessageEvent.Error(message = message.message)))
                 }
             }
         }
     }
 
     fun editMessage(id: Long, content: String) {
+        if (!isConnected()) return
         val editMessage = EditMessage(content = content)
         viewModelScope.launch {
             val wsId = channel.value?.workspaceId ?: return@launch
             val channelId = channel.value?.channelId ?: return@launch
             when (val response = repository.editMessage(editMessage, id, wsId, channelId).first()) {
                 is ResponseResult.Success -> {
-                    _messageEvent.value = Event(MessageEventS.Created())
+                    _messageEvent.value = Event(MessageEvent.Created())
                     modifyChannel()
                 }
                 is ResponseResult.Error -> {
-                    _messageEvent.postValue(Event(MessageEventS.Error(message = response.message)))
+                    _messageEvent.postValue(Event(MessageEvent.Error(message = response.message)))
                 }
             }
         }
@@ -104,17 +119,18 @@ class MessagesViewModel @ViewModelInject constructor(private val repository: Mes
     }
 
     fun deleteMessage(messageId: Long) {
+        if (!isConnected()) return
         viewModelScope.launch {
             val wsId = channel.value?.workspaceId ?: return@launch
             val channelId = channel.value?.channelId ?: return@launch
 
             when (val response = repository.deleteMessage(messageId, wsId, channelId).first()) {
                 is ResponseResult.Success -> {
-                    _messageEvent.value = Event(MessageEventS.Deleted())
+                    _messageEvent.value = Event(MessageEvent.Deleted())
                     modifyChannel()
                 }
                 is ResponseResult.Error -> {
-                    _messageEvent.postValue(Event(MessageEventS.Error(message = response.message)))
+                    _messageEvent.postValue(Event(MessageEvent.Error(message = response.message)))
                 }
             }
         }
